Refactor serde for Alias and AttributeReference
